Minor Revision:
Mostly, testing the Git/GitHub configuration and functionality after
migrating the development environment to the new Thinkpad laptop.

Changes:
(1) TabBar: Noticed that the TabBar is highed than desired on smaller
screens [such as this laptop, compared to my desktop screen previously
used for development].  Changed the TabBar to dynamically allocate the
width of each tab (versus fixed sizes which truncated the contents).
This fixed the truncation problem (a secondary effect resultant from the
height problem), but extended the TabBar slightly beyond the width of
the screen (losing only part of the width of a Tab), which I find
acceptable for now and possibly even for production (although it will
likely not occur once the height problem is fixed).

Explicitly defined the padding and insert to 0, which did not make any
difference. Will leave further height tweaking to a future build.

(2) Fixed the missing image for the Planing Tab (07).  This tab is
different from most in that this tab has both the base page and the
Form.ui page and the image is in the Form.ui page.  Keeping this style
(& code) (for now) in case it helps down the road.