Alpha 0.106 - Minor Revision
Implemented these changes discussed with Mark:
(1) The Done button on Utility menu switches back to the Opertator Tab
(Page01).
(2) Power-Off Crane button brings me back to the splash screen (Page00).
It also disables the SwipeView and TabBar mechanisms. <DONE>
(3) Remove Tab-Bar from splash screen (unable to do so; instead, I
disable the SwipeView & TabBar mechanisms.) <DONE>-<MODIFIED>
(4) Add two buttons to splash screen: <DONE>-<HACK>
  -- added the Power On Camera and Power On Crane buttons
(4a) the Power-up button on the splash screen will go to [the Error Log
menu if there's an error (not for demo) or] the Operator menu upon
normal start-up (for demo) <DONE>
(5) Change the tumbler to be 3D.  Remove the horizontal green bars.
<NOT-DONE>
<==this hasn't been done yet, needs to be done for the next commit
(&version).
(6) Modify the MyButton Object so that the color can be defined on
start-up.  Also, add am method to change button color and text
dynamically. <DONE>-<HACK>
(7) fix the positioning of text in main box of Page10. <DONE>
(8) make button clicks mutable depending on the mute button's setting.
<DONE>
 --<== needs two more related changes:
 (8a) to increase the max volume of the button clicks <DONE>
 (8b) to change the slider volume to logarithmic (currently, its beeping
sound plateus in the latter third of the volume slider: try to find a
function that linearizes this). <DONE>

Known Issues:
====================
Changing the Swipewiew updates the TabBar, but not vice-versa.  The
binding has likely been broken.  Will fix in next revision.

Components Affected:
====================

main:
- SwipeView and TabBar's enable/disable status is now bound to
GlobalProperties.qml's cranePowered variable.
- When crane is powered off, the GUI's logic switches back to splash
screen and disables the TabBar & SwipeView.

UtilityDrawer:
- implemented non-linear volume for slider (using value^3.5); changed
limits of slider from 0-1 to 0.2-1.

SunkenRectangle:
The implementation of whole SunkenRectangle component was modified to
correct for the border width of the "sunken" effect.  This was corrected
for all affected components (clock, tumbler, text box, etc).

CornerClock:
Corrected to work with new SunkenRectangle.qml.  Text weight changed to
Font.Medium, font point size increased to 14-equivalent.

Page09Utility:
- The Done button on Utility Page (09) switches back to the Opertator
Tab (Page01).
- Power-Off Crane on Utility Page (09) switches back to the splash
screen (Page00).  It also disables the SwipeView and TabBar mechanisms
via powerOffCrane().
- The <Node List> area (Tumbler) was corrected for the new
SunkenRectangle.qml.
- The main description field was corrected for the new
SunkenRectangle.qml.
- The main description field's text boxes now anchor.fill's it's parent
rectangle (it wasn't doing this automatically), fixing position errors
as the screen is re-dimensioned.

GlobalSounds:
replaced 3 WAV files (for click, press & release) with new ones that
could achieve higher volumes (the previous ones could not, even when
amplified to the max with a sound editor).

CustomButton1:
Modified it so that colors of "glow/shadow effects" can be specified
(instead of hardcorded).  Crrrently a hack implementation: needs to be
properly implemented for future release.

CustomButton2:
New component implemented for the SplashScreens buttons, implementing
the functionality that was needed (behavior changes from CustomButton1)
but not the functiotnality that was desired (transparency, look&feel).
Currently a hack implementation: a brand-new button probably needs to be
designed from the ground-up for the splash screen (transparent to
background).