Partially working Tumbler on Page10_MenuForm
Basically, implemented the Tumber more cleanl on Page10.  I moved the
list of values into a qml file (NodeListModel.qml).  It is located in
the Components directory.

This became a workaround for the errors I was getting before (not
finding some identifiers (id referenes) for tumber properties which were
not allowed in a SwipeView and therefore had to be moved from the
Page10MenuForm into the Page10Menu file).

Known Problems:
- currently getting a similar error for the itemShader.  Will try to fix
this in next version.

Planned changes:
- implement the same 'simplification' for the tumbler on Page09 -
Utility (in case it may be creating some conflicts).