Interim Commit - before discarding code
I made some changes trying to get OSCpack to work (before finding the
correct Ubuntu package).

I this will check-in my current code before deletion/changing begins.  I
do not intend to overwrite this commit in the future, just add further
commits: there were some good lines that might be useful as a reference
in the future.

The next commit should be an endpoint, closing the current code (with a
working keyboard and tumblers) BEFORE I begin to make changes due to
OSCpack.