Added Keyboard & Tumbler to Menu Page
Menu page has one Tumbler and one TextInput field (with working
dependent QtVirtualKeyboard).

Known limitations: the current implementation of QtVirtualKeyboard pops-
up and disappears automatically (and this is perfect) but it can occlude
the field when it pops up.  Need to discuss preferences with Mark (can
either slide the page up until the field is present [and this is the
standard accepted best-practice], or open a pop-up with only the
Keyboard and that Field [Qlarity-Style]).

Known Issues: The shader for the Tumbler on this page is giving an
Error.  At this time, I believe that is is a bug that needs to be
documented - I cannot find any other cause, having re-written the code
multiple times.