Added scrollview
Wrapped ScrollView around SwipeView (mainSwipeView) [our main UI
pardigm's backbone].  SwipeView is currently working, with nothing
broken the way that ScrollView.

Saving that in the current committ before making further changes.

TO-DO:
Next step is to implement the logic to automatically scroll the window
up as necessary when the virtual keyboard pops up and occludes the text
field.  This will likely be implemented only for text fields, and maybe
not all of them.

Number fields (and maybe some text fields) will likely be better served
by a top level pop-up with the proper keyboard and a field for the
entries.  Will, of course, get Mark's input on desired behavior &
choices.

Note: I have skipped all the TO-DOs from the previous commit.  Skipping
those for now and focusing on the (Auto?)Scrolling functionality that
may be required for the VirtualKeyboard.