Virtual Keyboard implemented (qwerty only)
Fully Implemented the Virtual Keyboard handling logic for the qwerty
keyboard.
- The keyboard slides up from the bottom
- The working screen slides up/down as needed for the focused field to be
just above the keyboard
- Moving from field to field functioning properly
- Good architecture used.

TO-DO:
- implement the various numeric keyboards (simpler to do, full-screen
pop-ups, the code prototype is already written and functional)
- [Next functionality] Implement the OSC calls and replies.
- Merge this back to the main trunk, then branch for OSC coding.

main.qml:
- implemented InputPanel (keyboard slides up) right off the main window.
- implemented ScrollView encapsulating the main SwipView, in order to
move the window to position the (focused) field just above the keyboard.

GlobalProperties:
- added two properties to better work with the ScrollView for the
VirtualKeyboard:
  + property real inputFieldHeight   // Holds the number of pixels
between the bottom of the TextInput (or equivalent) and the bottom of
the rootWindow.
  + property bool qwertyKBActive  // This controls whether the main
ScrollView (for qwerty KB) is active or not.  Should be set to 'true'
whenever a qwerty Keyboard is active.  Currently, my architecture is
that the field with the focus sets this; may need to evolve.  Logic for
this needs to be implemented when the logic for the other keyboards is
implemented.

Page10Menu:
- Implemented Bindings (one for each TextInput field with a qwerty
Keyboard - currently 3) that passesss the height of the current field for
the keyboard (i.e. top, main) ScrollView to work.

Page10MenuForm:
- Implemented implemented some minor changes to support the Bindings on
Page10Menu.